# ROX-12503 implement sensor's pipeline component