--- a/pkg/env/sensor.go
+++ b/pkg/env/sensor.go
@@ -18,4 +18,7 @@ var (
 
 	// LocalImageScanningEnabled is used to specify if Sensor should attempt to scan images via a local Scanner.
 	LocalImageScanningEnabled = RegisterBooleanSetting("ROX_LOCAL_IMAGE_SCANNING_ENABLED", false)
+
+	// EventPipelineOutputQueueSize is used to specify the size of the eventPipeline's output queue
+	EventPipelineOutputQueueSize = RegisterIntegerSetting("ROX_EVENT_PIPELINE_OUTPUT_QUEUE_SIZE", 100)
 )