--- a/sensor/kubernetes/eventpipeline/component/component.go
+++ b/sensor/kubernetes/eventpipeline/component/component.go
@@ -10,6 +10,7 @@ type PipelineComponent interface {
 }
 
 // OutputQueue component that redirects Resource Events and Alerts to the output channel
+//go:generate mockgen-wrapper
 type OutputQueue interface {
 	PipelineComponent
 	Send(detectionObject *ResourceEvent)