--- a/sensor/kubernetes/eventpipeline/output/output_impl.go
+++ b/sensor/kubernetes/eventpipeline/output/output_impl.go
@@ -2,7 +2,6 @@ package output
 
 import (
 	"github.com/stackrox/rox/generated/internalapi/central"
-	"github.com/stackrox/rox/pkg/concurrency"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/sensor/common/detector"
 	"github.com/stackrox/rox/sensor/kubernetes/eventpipeline/component"
@@ -14,7 +13,6 @@ var (
 
 type outputQueueImpl struct {
 	innerQueue   chan *component.ResourceEvent
-	stopSig      concurrency.Signal
 	forwardQueue chan *central.MsgFromSensor
 	detector     detector.Detector
 }
@@ -39,32 +37,27 @@ func (q *outputQueueImpl) Start() error {
 func (q *outputQueueImpl) Stop(_ error) {
 	defer close(q.innerQueue)
 	defer close(q.forwardQueue)
-	q.stopSig.Signal()
 }
 
 // runOutputQueue reads messages from the inner queue, forwards them to the forwardQueue channel
 // and sends the deployments (if needed) to Detector
 func (q *outputQueueImpl) runOutputQueue() {
 	for {
-		select {
-		case <-q.stopSig.Done():
+		msg, more := <-q.innerQueue
+		if !more {
 			return
-		case msg, more := <-q.innerQueue:
-			if !more {
-				return
+		}
+		for _, resourceUpdates := range msg.ForwardMessages {
+			q.forwardQueue <- &central.MsgFromSensor{
+				Msg: &central.MsgFromSensor_Event{
+					Event: resourceUpdates,
+				},
 			}
-			for _, resourceUpdates := range msg.ForwardMessages {
-				q.forwardQueue <- &central.MsgFromSensor{
-					Msg: &central.MsgFromSensor_Event{
-						Event: resourceUpdates,
-					},
-				}
-			}
+		}
 
-			q.detector.ReprocessDeployments(msg.CompatibilityReprocessDeployments...)
-			for _, detectorRequest := range msg.CompatibilityDetectionDeployment {
-				q.detector.ProcessDeployment(detectorRequest.Object, detectorRequest.Action)
-			}
+		q.detector.ReprocessDeployments(msg.CompatibilityReprocessDeployments...)
+		for _, detectorRequest := range msg.CompatibilityDetectionDeployment {
+			q.detector.ProcessDeployment(detectorRequest.Object, detectorRequest.Action)
 		}
 	}
 }