--- a/sensor/kubernetes/listener/resource_event_handler_impl.go
+++ b/sensor/kubernetes/listener/resource_event_handler_impl.go
@@ -7,6 +7,7 @@ import (
 	"github.com/stackrox/rox/pkg/kubernetes"
 	"github.com/stackrox/rox/pkg/sync"
 	"github.com/stackrox/rox/pkg/utils"
+	"github.com/stackrox/rox/sensor/kubernetes/eventpipeline/component"
 	"github.com/stackrox/rox/sensor/kubernetes/listener/resources"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -19,7 +20,7 @@ type resourceEventHandlerImpl struct {
 	eventLock  *sync.Mutex
 	dispatcher resources.Dispatcher
 
-	output           chan<- *central.MsgFromSensor
+	outputQueue      component.OutputQueue
 	syncingResources *concurrency.Flag
 
 	syncLock                   sync.Mutex
@@ -104,18 +105,8 @@ func (h *resourceEventHandlerImpl) sendResourceEvent(obj, oldObj interface{}, ac
 		kubernetes.TrimAnnotations(metaObj)
 	}
 
-	evWraps := h.dispatcher.ProcessEvent(obj, oldObj, action)
-	h.sendEvents(evWraps...)
-}
-
-func (h *resourceEventHandlerImpl) sendEvents(evWraps ...*central.SensorEvent) {
-	for _, evWrap := range evWraps {
-		h.output <- &central.MsgFromSensor{
-			Msg: &central.MsgFromSensor_Event{
-				Event: evWrap,
-			},
-		}
-	}
+	message := h.dispatcher.ProcessEvent(obj, oldObj, action)
+	h.outputQueue.Send(message)
 }
 
 func getObjUID(newObj interface{}) types.UID {