--- a/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorprofiles.go
+++ b/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorprofiles.go
@@ -60,5 +60,5 @@ func (c *ProfileDispatcher) ProcessEvent(obj, _ interface{}, action central.Reso
 			},
 		},
 	}
-	return component.WrapOutputMessage(events, nil, nil)
+	return component.NewResourceEvent(events, nil, nil)
 }