--- a/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorresults.go
+++ b/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorresults.go
@@ -73,5 +73,5 @@ func (c *ResultDispatcher) ProcessEvent(obj, _ interface{}, action central.Resou
 			},
 		},
 	}
-	return component.WrapOutputMessage(events, nil, nil)
+	return component.NewResourceEvent(events, nil, nil)
 }