--- a/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorresults.go
+++ b/sensor/kubernetes/listener/resources/complianceoperator/dispatchers/complianceoperatorresults.go
@@ -4,6 +4,7 @@ import (
 	"github.com/stackrox/rox/generated/internalapi/central"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/complianceoperator/api/v1alpha1"
+	"github.com/stackrox/rox/sensor/kubernetes/eventpipeline/component"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 )
@@ -39,7 +40,7 @@ func statusToProtoStatus(status v1alpha1.ComplianceCheckStatus) storage.Complian
 }
 
 // ProcessEvent processes a compliance operator check result
-func (c *ResultDispatcher) ProcessEvent(obj, _ interface{}, action central.ResourceAction) []*central.SensorEvent {
+func (c *ResultDispatcher) ProcessEvent(obj, _ interface{}, action central.ResourceAction) *component.ResourceEvent {
 	var complianceCheckResult v1alpha1.ComplianceCheckResult
 
 	unstructuredObject, ok := obj.(*unstructured.Unstructured)
@@ -54,7 +55,7 @@ func (c *ResultDispatcher) ProcessEvent(obj, _ interface{}, action central.Resou
 	}
 
 	id := string(complianceCheckResult.UID)
-	return []*central.SensorEvent{
+	events := []*central.SensorEvent{
 		{
 			Id:     id,
 			Action: action,
@@ -72,4 +73,5 @@ func (c *ResultDispatcher) ProcessEvent(obj, _ interface{}, action central.Resou
 			},
 		},
 	}
+	return component.WrapOutputMessage(events, nil, nil)
 }