## Description

Introducing a new sensor component that will wrap the k8s events management. This PR tackles all the refactoring on the handlers to use a queue to publish events and detector requests.

## Checklist
- [x] Investigated and inspected CI test results
- [x] Unit test and regression tests added
- ~[ ] Evaluated and added CHANGELOG entry if required~
- ~[ ] Determined and documented upgrade steps~
- ~[ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))~

If any of these don't apply, please comment below.

## Testing Performed

* CI
* Manually executing `sensor-integration-test` and `go-unit-tests`
* comparing benchmarks ([see](https://github.com/stackrox/stackrox/pull/3653#issuecomment-1300680692))
* long-running cluster with the following workload:
```yaml
deploymentWorkload:
- deploymentType: Deployment
  lifecycleDuration: 30m0s
  numDeployments: 1000
  numLifecycles: 0
  numLabels: 10
  randomLabels: true
  podWorkload:
    containerWorkload:
      numImages: 0
    lifecycleDuration: 10m0s
    numContainers: 3
    numPods: 5
    processWorkload:
      alertRate: 0.001
      processInterval: 1m0s
  updateInterval: 10m0s
networkPolicyWorkload:
- lifecycleDuration: 30m0s
  numNetworkPolicies: 500
  numLifecycles: 0
  updateInterval: 10m0s
networkWorkload:
  batchSize: 100
  flowInterval: 1s
nodeWorkload:
  numNodes: 1000
rbacWorkload:
  numBindings: 1000
  numRoles: 1000
  numServiceAccounts: 1000
serviceWorkload:
  numClusterIPs: 334
  numNodePorts: 333
  numLoadBalancers: 333
matchLabels: true
numNamespaces: 10
```