@fredrb 2022-11-02
For the record if anyone finds this PR in the future. This was first implemented in pair in this PR: #3495. All initial review comments are there.