@fredrb 2022-11-02
I ran the code in this branch against the benchmark in #3656 and compared the results with the benchmark shown in the PR:

```
% benchstat ./old.result.txt new-pipeline.result.txt
name          old time/op  new time/op  delta
_Pipeline-12   297ms ± 7%   267ms ± 8%  -9.88%  (p=0.032 n=5+5)
```

The pipeline here seems to be 9.8% faster. I think it's negligible since we're only considering deployments. But at least we know it's not significantly worse.