@fredrb 2022-11-02
/test openshift-oldest-operator-e2e-tests
/test openshift-newest-operator-e2e-tests